focus on tfd_dev
The problem is too big and complicated to start from the base. Rather, I'm going to start with the tfd api and then work backwards.

The new core concept is to have three "stages" (to be distinguished by directories).
stage 1 is /raw, this is data pulled from <some_source>(local, memory, storage)
stage 2 is /records, this is an intermediate/final step that converts the raw data into tfrecords
stage 3 is /preprocessed, the naming here isn't idea, but the idea is to have the data be analysis ready i.e. contains only relevant features and is also in tfrecord format.

when pulling from the tfd API, the data is going to be pulled into stage 3 (which is currently named /preprocessed.