set root_dir
This allows for the ability to control where all data should be stored. tf datasets are given a special prefix to help distinguish them as prebuilt.

The idea is to be able to create one dir containing all data for a given problem then use a Split to create the datasets from the directory